clippy fixes

+++ libafl/src/bolts/shmem.rs
@@ -356,6 +356,7 @@ pub mod unix_shmem {
     }
 
     /// Deinitialize this shmem instance
+    #[allow(clippy::clippy::clippy::unnecessary_cast)] // for c_ types
     unsafe fn unix_shmem_deinit(shm: *mut UnixShMem) {
         if shm.is_null() || (*shm).map.is_null() {
             /* Serialized map id */
@@ -369,6 +370,7 @@ pub mod unix_shmem {
 
     /// Functions to create Shared memory region, for observation channels and
     /// opening inputs and stuff.
+    #[allow(clippy::clippy::clippy::unnecessary_cast)] // for c_ types
     unsafe fn unix_shmem_init(shm: *mut UnixShMem, map_size: usize) -> *mut c_uchar {
         (*shm).map_size = map_size;
         (*shm).map = ptr::null_mut();
@@ -394,13 +396,14 @@ pub mod unix_shmem {
         if (*shm).map == -(1 as c_int) as *mut c_void as *mut c_uchar || (*shm).map.is_null() {
             shmctl((*shm).shm_id, 0 as c_int, ptr::null_mut());
             (*shm).shm_id = -(1 as c_int);
-            (*shm).shm_str[0 as c_int as usize] = 0u8;
+            (*shm).shm_str[0] = 0u8;
             return ptr::null_mut();
         }
         (*shm).map
     }
 
     /// Uses a shmap id string to open a shared map
+    #[allow(clippy::clippy::unnecessary_cast)] // for c_int and c_long
     unsafe fn unix_shmem_by_str(
         shm: *mut UnixShMem,
         shm_str: &CStr,
@@ -414,7 +417,7 @@ pub mod unix_shmem {
         strncpy(
             (*shm).shm_str.as_mut_ptr() as *mut c_char,
             shm_str.as_ptr() as *const c_char,
-            (size_of::<[c_char; 20]>() as c_ulong).wrapping_sub(1 as c_int as c_ulong),
+            (size_of::<[c_char; 20]>() as c_ulong).wrapping_sub(1 as c_ulong),
         );
         (*shm).shm_id = shm_str
             .to_str()
